Another go at fixing launching of cmd-key-happy via launchd on Mavericks
Complementary commit is 65976a6ccfe1670997521ad5536bb6f21e3274bf.  That
commit did allow cmd-key-happy to work when restarted, but not when the
machine was rebooted.  Calling AXAPIEnabled() appears to be very
necessary if the machine has rebooted, so sticking with this call until
an alternative truly becomes necessary.

Really really trying to close issue #12.