Use luaL_openlibs() to load libraries
Turns out my attempt to load a sandboxed version of libraries doesn't
actually work. The reason why this hasn't been an issue in the past is I
have only used print() in the cmd-key-happy.lua file which is/has always
been available. Trying to use io.write(), for example, fails as it is
not available in the Lua context.

I originally chose to try and load a sandboxed set of libraries to
mitigate any damage that could be done by loading ~/.cmd-key-happy.lua.
If that file is compromised then it could do... anything.